% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trampoline.R
\name{trambopoline}
\alias{trambopoline}
\title{Make a trampoline.}
\usage{
trambopoline(...)
}
\arguments{
\item{...}{A named list of generator functions. Named arguments are generator function
definitions where the name of the argument should be the desired name of the function
(that is referred to also within the function for recursion, see examples to get a clearer
idea of what this means). Passing multiple named arguments is possible and
allows specification of functions that can be used within the generator function that is
called in \code{call} (again, the examples might make this clearer).}
}
\value{
If \code{\link[=trm_return]{trm_return()}} or \code{\link[=trm_tailcall]{trm_tailcall()}} is called within the recursive generator
function, \code{trampoline()} will return the final return value from the final recursion.
Otherwise it will return \code{NULL} invisibly (in case the recursion is only for its
side-effects). See the examples for how this works.
}
\description{
This function takes a call to a generator factory, created by \code{\link[coro:generator]{coro::generator()}} and
runs it as a trampoline, which allows any recursion in the generator function to
recurse theoretically forever (but usually just more than can be handled by R's
default call stack limits).
}
