% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trampoline.R
\name{trm_return}
\alias{trm_return}
\title{Flag a return value}
\usage{
trm_return(x)
}
\arguments{
\item{x}{A value to be returned at the end of all recursions}
}
\value{
\code{x} with added class attribute 'trampoline_return'
}
\description{
Flag a return value
}
