% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trampoline.R
\name{trm_tailcall}
\alias{trm_tailcall}
\title{Flag a tail call}
\usage{
trm_tailcall(x)
}
\arguments{
\item{x}{A recursive call within generator fed to \code{\link[=trampoline]{trampoline()}}}
}
\value{
\code{x} with added class attribute 'trampoline_tailcall'
}
\description{
Flag a tail call
}
